floodlook_django/floodlook_django/scraper/requirements.txt
#pkg_resources==0.0.0
psycopg2-binary==2.9.2
pytz==2021.3
xmltodict==0.12.0
